Initial version
The Flash It Out app is intended to help learners practice words and phrases with flashcards showing a description and then revealing the word/phrase. The user can then mark a word/phrase as known or unknown so that the app can prioritize unknown words and show them more often, as well as the least practiced words.